% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compomon-package.R
\docType{package}
\name{compomon-package}
\alias{compomon}
\alias{compomon-package}
\title{compomon: MULTIVARIATE STATISTICAL PROCESS MONITORING of COMPOAITIONAL DATA}
\description{
Computes the biplot properties proposed by Aitchison & Greenacre (2002) and calculates monitoring statistics and variable contributions for log-transformed compositions in PCA-reduced dimensions.
}
\author{
\strong{Maintainer}: Jennifer Liebenberg \email{jenniferleighconn@gmail.com} (\href{https://orcid.org/0000-0002-3811-4957}{ORCID})

}
\keyword{internal}
